Returned frames as a raw []byte
For transmitting frames between processes (which we're starting to do
now) it's useful to have frames as raw byte slices instead of a 2D
array of ints. The new RawFrame type represents a raw frame and has a
method to allow conversion to the more useful Frame type.

Also renamed the internal `frame` type to `frameBuilder` to avoid
confusion.